Springboot/workspace-spring-tool-suite-4-4.18.1.RELEASE/CalcBoot/src/main/java/Advajavaassesment/CharacterCounter.java
package Advajavaassesment;

import java.util.HashMap;
import java.util.Map;

public class CharacterCounter {
    public static Map<Character, Integer> countCharacters(String input) {
        Map<Character, Integer> charCount = new HashMap<>();

        for (char ch : input.toCharArray()) {
            charCount.put(ch, charCount.getOrDefault(ch, 0) + 1);
        }

        return charCount;
    }

    public static void main(String[] args) {
        String input = "aba";
        Map<Character, Integer> result = countCharacters(input);
        System.out.println(result);
    }
}


